requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
cfgv==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distlib==0.3.9
Django==5.1.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
drf-nested-routers==0.94.1
drf-spectacular==0.28.0
filelock==3.16.1
flake8==7.1.1
identify==2.6.1
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
mccabe==0.7.0
nodeenv==1.9.1
platformdirs==4.3.6
pre_commit==4.0.1
prompt_toolkit==3.0.48
pycodestyle==2.12.1
pyflakes==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.1.1
referencing==0.35.1
rpds-py==0.21.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
vine==5.1.0
virtualenv==20.27.0
wcwidth==0.2.13
